A class in which the number of hidden layers can be varied
A class which allows one to vary the number of nodes in a hidden layer and analyze it's effectiveness based on the error generated.

While creating it's object one must specify the number of hidden layers as a parameter.

Eg- neural_network = NeuralNetwork(3)

By default it has a hidden layer of 2 nodes.